Introduce drone location changed event
New event is fired when location message from drone accepted and stored,
and this is not the first message from the drone. Parameters of this
event are old and new locations. The event is published by location
received handler.

Location received handler tests were extended to cover new event
functionality.